shader/analyzer: Add function to extract predicate
