+++ src/emulator/shader/CMakeLists.txt
@@ -7,6 +7,7 @@ add_library(
 	include/shader/matcher.h
 	include/shader/decoder_detail.h
 	include/shader/usse_disasm.h
+	include/shader/usse_program_analyzer.h
 	include/shader/usse_decoder_helpers.h
 	include/shader/usse_translator.h
 	include/shader/usse_translator_entry.h
@@ -20,6 +21,7 @@ add_library(
 	src/translator/texture.cpp
 	src/translator/utils.cpp
 	src/usse_disasm.cpp
+	src/usse_program_analyzer.cpp
 	src/usse_decode_helpers.cpp
 	src/usse_translator_entry.cpp
 	src/spirv_recompiler.cpp
